refactor: move ioctl-related code and rename functions
