+++ src/channel.rs
@@ -10,7 +10,7 @@ use std::{
 
 use libc::{c_int, c_void, size_t};
 use crate::reply::ReplySender;
-use crate::ll::ioctl::{deregister_backing_id, register_backing_id};
+use crate::ll::ioctl::{ioctl_close_backing, ioctl_open_backing};
 
 
 /// A raw communication channel to the FUSE kernel driver
@@ -123,12 +123,12 @@ impl ChannelSender {
     /// Registers a file descriptor for passthrough and returns a backing ID.
     #[cfg(feature = "abi-7-40")]
     pub fn open_backing(&self, fd: u32) -> std::io::Result<u32> {
-        register_backing_id(&self.0, fd)
+        ioctl_open_backing(&self.0, fd)
     }
 
     /// Deregisters a backing ID.
     #[cfg(feature = "abi-7-40")]
     pub fn close_backing(&self, backing_id: u32) -> std::io::Result<u32> {
-        deregister_backing_id(&self.0, backing_id)
+        ioctl_close_backing(&self.0, backing_id)
     }
 }
