+++ src/ll/fuse_abi.rs
@@ -1175,35 +1175,3 @@ pub struct fuse_copy_file_range_in {
     pub flags: u64,
 }
 
-// ----- Ioctl for Passthrough -----
-
-/// The `fuse_backing_map_out` struct is used to pass information about a backing file
-/// descriptor to the kernel.
-#[repr(C)]
-pub struct fuse_backing_map_out {
-    pub fd: u32,
-    pub flags: u32,
-    pub padding: u64,
-}
-
-const FUSE_DEV_IOC_MAGIC: u8 = 229;
-const FUSE_DEV_IOC_BACKING_OPEN: u8 = 1;
-const FUSE_DEV_IOC_BACKING_CLOSE: u8 = 2;
-
-/// This ioctl is used to register a backing file descriptor with the kernel.
-/// The kernel will return a backing ID that can be used to refer to the file descriptor in
-/// subsequent operations.
-nix::ioctl_write_ptr!(
-    fuse_dev_ioc_backing_open,
-    FUSE_DEV_IOC_MAGIC,
-    FUSE_DEV_IOC_BACKING_OPEN,
-    fuse_backing_map_out
-);
-
-/// This ioctl is used to deregister a backing file descriptor.
-nix::ioctl_write_ptr!(
-    fuse_dev_ioc_backing_close,
-    FUSE_DEV_IOC_MAGIC,
-    FUSE_DEV_IOC_BACKING_CLOSE,
-    u32
-);
